Implemented UI features
Added health functionality - Collision detection between the player and the Gorilla. When they collide the player loses 1 health point and becomes invulnerable for 2 seconds. Once the player hits 0 HP, a message is displayed in the console saying, "You Died!", and the player movement is set to 0.

Added oxygen functionality - Checks when the oxygen status is red. When it is, the oxygen timer starts. The player has 60 seconds to stabilize the oxygen node otherwise they instantaneously die, printing the same "You Died!" message to the console, and setting the player movement to 0.

Added a "Ship Status" area in the top left - Does not currently function, simply a placeholder to continue working on in the future.